refactor: split all modules over 600 lines into focused sub-modules
Enforce the 600-line ceiling across five files using SOLID/DRY/KISS:

transaction/openapi.py (714->12 lines as re-export hub)
  - openapi_shared.py    (102) -- constants, examples, shared helpers
  - openapi_mutations.py (257) -- write-operation OpenAPI docs
  - openapi_queries.py   (448) -- read-operation OpenAPI docs

transaction/report_resources.py (783->44 lines as re-export hub)
  - list_resources.py      (279) -- collection and list endpoints
  - detail_resources.py    (148) -- detail, restore, force-delete endpoints
  - analytics_resources.py (432) -- summary, dashboard, period endpoints

middleware/rate_limit.py (644->520 lines)
  - rate_limit_storage.py (138) -- RateLimitStorage protocol + backends

controllers/subscription_controller.py (692->395 lines)
  - subscription_webhook_payload.py (322) -- pure payload parsing helpers

application/services/transaction_ledger_service.py (828->579 lines)
  - transaction/errors.py      (14)  -- TransactionApplicationError dataclass
  - transaction/validators.py  (182) -- validators and field normalizers
  - transaction/query_helpers.py (178) -- ordering, update, serialization

All backward-compatible: existing import paths preserved via re-exports.
Monkeypatch path in test updated to match new module location.
1039 tests pass, ruff + mypy clean.
