+++ app/application/services/transaction/__init__.py
@@ -1,0 +1,1 @@
+"""Transaction sub-package — domain building blocks for the ledger service."""
