+++ app/controllers/transaction/report_resources.py
@@ -1,770 +1,32 @@
+"""Backward-compatible re-exports for transaction report resources.
+
+All resource classes and legacy symbols re-exported to keep routes.py,
+tests, and other callers working without modification.
+"""
+
 from __future__ import annotations
 
-from collections.abc import Callable
-from typing import Any
-from uuid import UUID
-
-from flask import Response, request
-from flask_apispec.views import MethodResource
-
-from app.application.services.public_error_mapper_service import (
-    map_validation_exception,
+from app.controllers.transaction.analytics_resources import (
+    TransactionDeletedResource,
+    TransactionDuePeriodResource,
+    TransactionExpensePeriodResource,
+    TransactionMonthlyDashboardResource,
+    TransactionSummaryResource,
 )
-from app.application.services.transaction_application_service import (
-    TransactionApplicationError,
+from app.controllers.transaction.detail_resources import (
+    TransactionDetailResource,
+    TransactionForceDeleteResource,
+    TransactionRestoreResource,
 )
-from app.auth import current_user_id
-from app.extensions.database import db
-from app.models.transaction import Transaction, TransactionStatus, TransactionType
-from app.services.transaction_analytics_service import TransactionAnalyticsService
-from app.utils.typed_decorators import typed_doc as doc
-from app.utils.typed_decorators import typed_jwt_required as jwt_required
-
-from .dependencies import get_transaction_dependencies
-from .openapi import (
-    TRANSACTION_ACTIVE_LIST_DOC,
-    TRANSACTION_ACTIVE_LIST_LEGACY_DOC,
-    TRANSACTION_DASHBOARD_DOC,
-    TRANSACTION_DELETED_LIST_DOC,
-    TRANSACTION_DETAIL_DOC,
-    TRANSACTION_DUE_PERIOD_DOC,
-    TRANSACTION_EXPENSE_PERIOD_DOC,
-    TRANSACTION_FORCE_DELETE_DOC,
-    TRANSACTION_RESTORE_DOC,
-    TRANSACTION_SUMMARY_DOC,
-)
-from .utils import (
-    _apply_deprecation_headers,
-    _compat_error,
-    _compat_success,
+from app.controllers.transaction.list_resources import (
+    TransactionCollectionResource,
+    TransactionListActiveResource,
     _guard_revoked_token,
-    _internal_error_response,
-    _parse_optional_date,
-    _parse_optional_uuid,
-    _parse_positive_int,
     _resolve_transaction_ordering,
 )
+from app.services.transaction_analytics_service import TransactionAnalyticsService
 
-# Keep symbol available as monkeypatch target for legacy tests.
 _LEGACY_ANALYTICS_SERVICE = TransactionAnalyticsService
-
-
-def _validation_error_response(
-    *,
-    exc: Exception,
-    fallback_message: str,
-) -> Response:
-    mapped_error = map_validation_exception(exc, fallback_message=fallback_message)
-    return _compat_error(
-        legacy_payload={"error": mapped_error.message},
-        status_code=mapped_error.status_code,
-        message=mapped_error.message,
-        error_code=mapped_error.code,
-        details=mapped_error.details,
-    )
-
-
-def _first_query_value(*names: str) -> str | None:
-    for name in names:
-        value = request.args.get(name)
-        if value is not None:
-            return str(value)
-    return None
-
-
-def _parse_active_list_query_params() -> dict[str, Any]:
-    return {
-        "page": _parse_positive_int(
-            request.args.get("page"), default=1, field_name="page"
-        ),
-        "per_page": _parse_positive_int(
-            request.args.get("per_page"),
-            default=10,
-            field_name="per_page",
-        ),
-        "transaction_type": request.args.get("type"),
-        "status": request.args.get("status"),
-        "start_date": _parse_optional_date(
-            request.args.get("start_date"), "start_date"
-        ),
-        "end_date": _parse_optional_date(request.args.get("end_date"), "end_date"),
-        "tag_id": _parse_optional_uuid(request.args.get("tag_id"), "tag_id"),
-        "account_id": _parse_optional_uuid(
-            request.args.get("account_id"), "account_id"
-        ),
-        "credit_card_id": _parse_optional_uuid(
-            request.args.get("credit_card_id"), "credit_card_id"
-        ),
-    }
-
-
-def _active_list_date_range_error(*, start_date: Any, end_date: Any) -> Response | None:
-    if not start_date or not end_date or start_date <= end_date:
-        return None
-    message = "Parâmetro 'start_date' não pode ser maior que 'end_date'."
-    return _compat_error(
-        legacy_payload={"error": message},
-        status_code=400,
-        message=message,
-        error_code="VALIDATION_ERROR",
-    )
-
-
-def _apply_active_list_type_filter(
-    query: Any, transaction_type: str | None
-) -> tuple[Any, Response | None]:
-    if not transaction_type:
-        return query, None
-    try:
-        return (
-            query.filter(Transaction.type == TransactionType(transaction_type.lower())),
-            None,
-        )
-    except ValueError:
-        message = "Parâmetro 'type' inválido. Use 'income' ou 'expense'."
-        return query, _compat_error(
-            legacy_payload={"error": message},
-            status_code=400,
-            message=message,
-            error_code="VALIDATION_ERROR",
-        )
-
-
-def _apply_active_list_status_filter(
-    query: Any, status: str | None
-) -> tuple[Any, Response | None]:
-    if not status:
-        return query, None
-    try:
-        return (
-            query.filter(Transaction.status == TransactionStatus(status.lower())),
-            None,
-        )
-    except ValueError:
-        message = (
-            "Parâmetro 'status' inválido. "
-            "Use paid, pending, cancelled, postponed ou overdue."
-        )
-        return query, _compat_error(
-            legacy_payload={"error": message},
-            status_code=400,
-            message=message,
-            error_code="VALIDATION_ERROR",
-        )
-
-
-def _apply_active_list_optional_filters(
-    query: Any,
-    *,
-    start_date: Any,
-    end_date: Any,
-    tag_id: Any,
-    account_id: Any,
-    credit_card_id: Any,
-) -> Any:
-    if start_date:
-        query = query.filter(Transaction.due_date >= start_date)
-    if end_date:
-        query = query.filter(Transaction.due_date <= end_date)
-    if tag_id:
-        query = query.filter(Transaction.tag_id == tag_id)
-    if account_id:
-        query = query.filter(Transaction.account_id == account_id)
-    if credit_card_id:
-        query = query.filter(Transaction.credit_card_id == credit_card_id)
-    return query
-
-
-def _build_active_list_response(user_uuid: UUID) -> Response:
-    params = _parse_active_list_query_params()
-    range_error = _active_list_date_range_error(
-        start_date=params["start_date"],
-        end_date=params["end_date"],
-    )
-    if range_error is not None:
-        return range_error
-
-    dependencies = get_transaction_dependencies()
-    query_service = dependencies.transaction_query_service_factory(user_uuid)
-    result = query_service.get_active_transactions(
-        page=params["page"],
-        per_page=params["per_page"],
-        transaction_type=params["transaction_type"],
-        status=params["status"],
-        start_date=params["start_date"],
-        end_date=params["end_date"],
-        tag_id=params["tag_id"],
-        account_id=params["account_id"],
-        credit_card_id=params["credit_card_id"],
-    )
-    pagination = result["pagination"]
-    serialized = result["items"]
-    return _compat_success(
-        legacy_payload={
-            "transactions": serialized,
-            "total": pagination["total"],
-            "page": pagination["page"],
-            "per_page": pagination["per_page"],
-        },
-        status_code=200,
-        message="Lista de transações ativas",
-        data={"transactions": serialized},
-        meta={"pagination": pagination},
-    )
-
-
-def _build_transaction_detail_response(
-    *,
-    user_uuid: UUID,
-    transaction_id: UUID,
-) -> Response:
-    dependencies = get_transaction_dependencies()
-    query_service = dependencies.transaction_query_service_factory(user_uuid)
-    serialized = query_service.get_transaction(transaction_id)
-    return _compat_success(
-        legacy_payload={"transaction": serialized},
-        status_code=200,
-        message="Detalhe da transação carregado com sucesso",
-        data={"transaction": serialized},
-    )
-
-
-def _handle_transaction_request(
-    *,
-    builder: Callable[[UUID], Response],
-    internal_error_message: str,
-    log_context: str,
-    validation_error_message: str | None = None,
-) -> Response:
-    token_error = _guard_revoked_token()
-    if token_error is not None:
-        return token_error
-
-    user_uuid = current_user_id()
-    try:
-        return builder(user_uuid)
-    except TransactionApplicationError as exc:
-        return _compat_error(
-            legacy_payload={"error": exc.message, "details": exc.details},
-            status_code=exc.status_code,
-            message=exc.message,
-            error_code=exc.code,
-            details=exc.details,
-        )
-    except ValueError as exc:
-        if validation_error_message is None:
-            raise
-        return _validation_error_response(
-            exc=exc,
-            fallback_message=validation_error_message,
-        )
-    except Exception:
-        db.session.rollback()
-        return _internal_error_response(
-            message=internal_error_message,
-            log_context=log_context,
-        )
-
-
-def _handle_active_transactions_request() -> Response:
-    return _handle_transaction_request(
-        builder=_build_active_list_response,
-        internal_error_message="Erro ao buscar transações ativas",
-        log_context="transaction.list_active_failed",
-        validation_error_message="Parâmetros de listagem inválidos.",
-    )
-
-
-def _handle_transaction_detail_request(*, transaction_id: UUID) -> Response:
-    return _handle_transaction_request(
-        builder=lambda user_uuid: _build_transaction_detail_response(
-            user_uuid=user_uuid,
-            transaction_id=transaction_id,
-        ),
-        internal_error_message="Erro ao buscar detalhe da transação",
-        log_context="transaction.detail_failed",
-    )
-
-
-def _parse_summary_pagination() -> tuple[int, int]:
-    page = _parse_positive_int(request.args.get("page"), default=1, field_name="page")
-    per_page = _parse_positive_int(
-        _first_query_value("per_page", "page_size"),
-        default=10,
-        field_name="per_page",
-    )
-    return page, per_page
-
-
-def _parse_expense_period_params() -> dict[str, Any]:
-    start_date = _parse_optional_date(
-        _first_query_value("start_date", "startDate"),
-        "start_date",
-    )
-    end_date = _parse_optional_date(
-        _first_query_value("end_date", "finalDate"),
-        "end_date",
-    )
-    return {
-        "start_date": start_date,
-        "end_date": end_date,
-        "page": _parse_positive_int(
-            request.args.get("page"), default=1, field_name="page"
-        ),
-        "per_page": _parse_positive_int(
-            request.args.get("per_page"), default=10, field_name="per_page"
-        ),
-        "order_by": str(request.args.get("order_by", "due_date")).strip().lower(),
-        "order": str(request.args.get("order", "desc")).strip().lower(),
-    }
-
-
-def _parse_due_range_params() -> dict[str, Any]:
-    return {
-        "start_date": _first_query_value("start_date", "initialDate"),
-        "end_date": _first_query_value("end_date", "finalDate"),
-        "page": _parse_positive_int(
-            request.args.get("page"), default=1, field_name="page"
-        ),
-        "per_page": _parse_positive_int(
-            request.args.get("per_page"), default=10, field_name="per_page"
-        ),
-        "order_by": str(request.args.get("order_by", "overdue_first")).strip().lower(),
-    }
-
-
-def _deprecated_expense_alias_response(response: Response) -> Response:
-    return _apply_deprecation_headers(
-        response,
-        successor_endpoint="/transactions?type=expense",
-        successor_method="GET",
-    )
-
-
-def _deprecated_dashboard_alias_response(response: Response) -> Response:
-    return _apply_deprecation_headers(
-        response,
-        successor_endpoint="/dashboard/overview",
-        successor_method="GET",
-    )
-
-
-class TransactionCollectionResource(MethodResource):
-    @doc(**TRANSACTION_ACTIVE_LIST_DOC)
-    @jwt_required()
-    def get(self) -> Response:
-        return _handle_active_transactions_request()
-
-
-class TransactionDetailResource(MethodResource):
-    @doc(**TRANSACTION_DETAIL_DOC)
-    @jwt_required()
-    def get(self, transaction_id: UUID) -> Response:
-        return _handle_transaction_detail_request(transaction_id=transaction_id)
-
-
-class TransactionSummaryResource(MethodResource):
-    @doc(**TRANSACTION_SUMMARY_DOC)
-    @jwt_required()
-    def get(self) -> Response:
-        token_error = _guard_revoked_token()
-        if token_error is not None:
-            return token_error
-
-        user_uuid = current_user_id()
-        try:
-            page, per_page = _parse_summary_pagination()
-            dependencies = get_transaction_dependencies()
-            query_service = dependencies.transaction_query_service_factory(user_uuid)
-            result = query_service.get_month_summary(
-                month=str(request.args.get("month", "")),
-                page=page,
-                per_page=per_page,
-            )
-            paginated = result["paginated"]
-
-            return _compat_success(
-                legacy_payload={
-                    "month": result["month"],
-                    "income_total": result["income_total"],
-                    "expense_total": result["expense_total"],
-                    **paginated,
-                },
-                status_code=200,
-                message="Resumo mensal calculado com sucesso",
-                data={
-                    "month": result["month"],
-                    "income_total": result["income_total"],
-                    "expense_total": result["expense_total"],
-                    "items": paginated["data"],
-                },
-                meta={
-                    "pagination": {
-                        "total": paginated["total"],
-                        "page": paginated["page"],
-                        "per_page": paginated["page_size"],
-                        "has_next_page": paginated["has_next_page"],
-                    }
-                },
-            )
-        except TransactionApplicationError as exc:
-            return _compat_error(
-                legacy_payload={"error": exc.message, "details": exc.details},
-                status_code=exc.status_code,
-                message=exc.message,
-                error_code=exc.code,
-                details=exc.details,
-            )
-        except ValueError as exc:
-            return _validation_error_response(
-                exc=exc,
-                fallback_message="Parâmetro de mês inválido.",
-            )
-        except Exception:
-            db.session.rollback()
-            return _internal_error_response(
-                message="Erro ao calcular resumo mensal",
-                log_context="transaction.monthly_summary_failed",
-            )
-
-
-class TransactionMonthlyDashboardResource(MethodResource):
-    @doc(**TRANSACTION_DASHBOARD_DOC)
-    @jwt_required()
-    def get(self) -> Response:
-        token_error = _guard_revoked_token()
-        if token_error is not None:
-            return token_error
-
-        user_uuid = current_user_id()
-        try:
-            dependencies = get_transaction_dependencies()
-            query_service = dependencies.transaction_query_service_factory(user_uuid)
-            result = query_service.get_dashboard_overview(
-                month=str(request.args.get("month", "")),
-            )
-
-            return _deprecated_dashboard_alias_response(
-                _compat_success(
-                    legacy_payload={
-                        "month": result["month"],
-                        "income_total": result["income_total"],
-                        "expense_total": result["expense_total"],
-                        "balance": result["balance"],
-                        "counts": result["counts"],
-                        "top_expense_categories": result["top_expense_categories"],
-                        "top_income_categories": result["top_income_categories"],
-                    },
-                    status_code=200,
-                    message="Dashboard mensal calculado com sucesso",
-                    data={
-                        "month": result["month"],
-                        "totals": {
-                            "income_total": result["income_total"],
-                            "expense_total": result["expense_total"],
-                            "balance": result["balance"],
-                        },
-                        "counts": result["counts"],
-                        "top_categories": {
-                            "expense": result["top_expense_categories"],
-                            "income": result["top_income_categories"],
-                        },
-                    },
-                )
-            )
-        except TransactionApplicationError as exc:
-            return _compat_error(
-                legacy_payload={"error": exc.message, "details": exc.details},
-                status_code=exc.status_code,
-                message=exc.message,
-                error_code=exc.code,
-                details=exc.details,
-            )
-        except ValueError as exc:
-            return _validation_error_response(
-                exc=exc,
-                fallback_message="Parâmetro de mês inválido.",
-            )
-        except Exception:
-            db.session.rollback()
-            return _internal_error_response(
-                message="Erro ao calcular dashboard mensal",
-                log_context="transaction.monthly_dashboard_failed",
-            )
-
-
-class TransactionForceDeleteResource(MethodResource):
-    @doc(**TRANSACTION_FORCE_DELETE_DOC)
-    @jwt_required()
-    def delete(self, transaction_id: UUID) -> Response:
-        token_error = _guard_revoked_token()
-        if token_error is not None:
-            return token_error
-
-        user_id = current_user_id()
-        transaction = Transaction.query.filter_by(
-            id=transaction_id, user_id=user_id, deleted=True
-        ).first()
-
-        if not transaction:
-            return _compat_error(
-                legacy_payload={
-                    "error": "Transação não encontrada ou não está deletada."
-                },
-                status_code=404,
-                message="Transação não encontrada ou não está deletada.",
-                error_code="NOT_FOUND",
-            )
-
-        try:
-            db.session.delete(transaction)
-            db.session.commit()
-            return _compat_success(
-                legacy_payload={"message": "Transação removida permanentemente."},
-                status_code=200,
-                message="Transação removida permanentemente.",
-                data={},
-            )
-        except Exception:
-            db.session.rollback()
-            return _internal_error_response(
-                message="Erro ao deletar permanentemente",
-                log_context="transaction.force_delete_failed",
-            )
-
-
-class TransactionExpensePeriodResource(MethodResource):
-    @doc(**TRANSACTION_EXPENSE_PERIOD_DOC)
-    @jwt_required()
-    def get(self) -> Response:
-        token_error = _guard_revoked_token()
-        if token_error is not None:
-            return token_error
-
-        user_uuid = current_user_id()
-
-        try:
-            params = _parse_expense_period_params()
-            start_date = params["start_date"]
-            end_date = params["end_date"]
-            if not start_date and not end_date:
-                missing_date_error = (
-                    "Informe ao menos um parâmetro: 'start_date' ou 'end_date'."
-                )
-                return _deprecated_expense_alias_response(
-                    _compat_error(
-                        legacy_payload={"error": missing_date_error},
-                        status_code=400,
-                        message=missing_date_error,
-                        error_code="VALIDATION_ERROR",
-                    )
-                )
-
-            if start_date and end_date and start_date > end_date:
-                invalid_range_error = (
-                    "Parâmetro 'start_date' não pode ser maior que 'end_date'."
-                )
-                return _deprecated_expense_alias_response(
-                    _compat_error(
-                        legacy_payload={"error": invalid_range_error},
-                        status_code=400,
-                        message=invalid_range_error,
-                        error_code="VALIDATION_ERROR",
-                    )
-                )
-
-            page = params["page"]
-            per_page = params["per_page"]
-            order_by = params["order_by"]
-            order = params["order"]
-            ordering_clause = _resolve_transaction_ordering(order_by, order)
-            dependencies = get_transaction_dependencies()
-            query_service = dependencies.transaction_query_service_factory(user_uuid)
-            result = query_service.get_expense_period(
-                start_date=start_date,
-                end_date=end_date,
-                page=page,
-                per_page=per_page,
-                ordering_clause=ordering_clause,
-            )
-            serialized_expenses = result["expenses"]
-            counts_payload = result["counts"]
-            pagination = result["pagination"]
-
-            response = _compat_success(
-                legacy_payload={
-                    "expenses": serialized_expenses,
-                    "total": pagination["total"],
-                    "page": page,
-                    "per_page": per_page,
-                    "counts": counts_payload,
-                },
-                status_code=200,
-                message="Lista de despesas por período",
-                data={"expenses": serialized_expenses, "counts": counts_payload},
-                meta={
-                    "pagination": {
-                        "total": pagination["total"],
-                        "page": page,
-                        "per_page": per_page,
-                        "pages": pagination["pages"],
-                    }
-                },
-            )
-            return _deprecated_expense_alias_response(response)
-        except ValueError as exc:
-            return _deprecated_expense_alias_response(
-                _validation_error_response(
-                    exc=exc,
-                    fallback_message="Parâmetros de período inválidos.",
-                )
-            )
-        except Exception:
-            db.session.rollback()
-            return _deprecated_expense_alias_response(
-                _internal_error_response(
-                    message="Erro ao buscar despesas por período",
-                    log_context="transaction.expenses_period_failed",
-                )
-            )
-
-
-class TransactionDeletedResource(MethodResource):
-    @doc(**TRANSACTION_DELETED_LIST_DOC)
-    @jwt_required()
-    def get(self) -> Response:
-        token_error = _guard_revoked_token()
-        if token_error is not None:
-            return token_error
-
-        user_uuid = current_user_id()
-        try:
-            dependencies = get_transaction_dependencies()
-            query_service = dependencies.transaction_query_service_factory(user_uuid)
-            serialized = query_service.list_deleted_transactions()
-            return _compat_success(
-                legacy_payload={"deleted_transactions": serialized},
-                status_code=200,
-                message="Lista de transações deletadas",
-                data={"deleted_transactions": serialized},
-                meta={"total": len(serialized)},
-            )
-        except Exception:
-            db.session.rollback()
-            return _internal_error_response(
-                message="Erro ao buscar transações deletadas",
-                log_context="transaction.list_deleted_failed",
-            )
-
-
-class TransactionDuePeriodResource(MethodResource):
-    @doc(**TRANSACTION_DUE_PERIOD_DOC)
-    @jwt_required()
-    def get(self) -> Response:
-        token_error = _guard_revoked_token()
-        if token_error is not None:
-            return token_error
-
-        user_uuid = current_user_id()
-        try:
-            params = _parse_due_range_params()
-
-            dependencies = get_transaction_dependencies()
-            query_service = dependencies.transaction_query_service_factory(user_uuid)
-            result = query_service.get_due_transactions(
-                start_date=params["start_date"],
-                end_date=params["end_date"],
-                page=params["page"],
-                per_page=params["per_page"],
-                order_by=params["order_by"],
-            )
-            pagination = result["pagination"]
-            counts = result["counts"]
-
-            return _compat_success(
-                legacy_payload={
-                    "transactions": result["items"],
-                    "total": pagination["total"],
-                    "page": pagination["page"],
-                    "per_page": pagination["per_page"],
-                    "counts": counts,
-                },
-                status_code=200,
-                message="Lista de vencimentos por período",
-                data={"transactions": result["items"], "counts": counts},
-                meta={
-                    "pagination": {
-                        "total": pagination["total"],
-                        "page": pagination["page"],
-                        "per_page": pagination["per_page"],
-                        "pages": pagination["pages"],
-                    }
-                },
-            )
-        except TransactionApplicationError as exc:
-            return _compat_error(
-                legacy_payload={"error": exc.message, "details": exc.details},
-                status_code=exc.status_code,
-                message=exc.message,
-                error_code=exc.code,
-                details=exc.details,
-            )
-        except ValueError as exc:
-            return _validation_error_response(
-                exc=exc,
-                fallback_message="Parâmetros de vencimentos inválidos.",
-            )
-        except Exception:
-            db.session.rollback()
-            return _internal_error_response(
-                message="Erro ao buscar vencimentos por período",
-                log_context="transaction.due_period_failed",
-            )
-
-
-class TransactionRestoreResource(MethodResource):
-    @doc(**TRANSACTION_RESTORE_DOC)
-    @jwt_required()
-    def patch(self, transaction_id: UUID) -> Response:
-        token_error = _guard_revoked_token()
-        if token_error is not None:
-            return token_error
-
-        user_uuid = current_user_id()
-        transaction = Transaction.query.filter_by(
-            id=transaction_id, user_id=user_uuid, deleted=True
-        ).first()
-        if transaction is None:
-            return _compat_error(
-                legacy_payload={"error": "Transação não encontrada."},
-                status_code=404,
-                message="Transação não encontrada.",
-                error_code="NOT_FOUND",
-            )
-
-        try:
-            transaction.deleted = False
-            db.session.commit()
-            return _compat_success(
-                legacy_payload={"message": "Transação restaurada com sucesso"},
-                status_code=200,
-                message="Transação restaurada com sucesso",
-                data={},
-            )
-        except Exception:
-            db.session.rollback()
-            return _internal_error_response(
-                message="Erro ao restaurar transação",
-                log_context="transaction.restore_failed",
-            )
-
-
-class TransactionListActiveResource(MethodResource):
-    @doc(**TRANSACTION_ACTIVE_LIST_LEGACY_DOC)
-    @jwt_required()
-    def get(self) -> Response:
-        return _handle_active_transactions_request()
-
 
 __all__ = [
     "TransactionCollectionResource",
